feat: refactor bank service
